Renamed, added more images, and finished html file
Renamed the picture of the jellyfish to jellyfish.jpg for a better file name. Added
more images to be added into the images folder and onto the html file. Images include
a duck, clouds and the ocean, a moon, and a whale.

The html file is completed for now until I need to add more classes when I style it in
css file. Added the images, quotes, buttons, and a footer. Will now move on to styling
the html file to make it look better.