Changed and added in html and styled the header
Had to add more classes in the html file to be able to style certain elements.
Styled the header first, the header is not done but will finish later. Centered the
entire body to be uniform with each other. Styled the buttons to look nicer. Removed
the bullet points after the links for a better look and removed the padding and margins
from the list.