Styled the quote and header section
Quote section should be fully styled, however, not sure if I should have used
the <i></i> element or styled the italics in. I opted to use the font-style italics.

The header section is still a bit tricky area for me. I decided to make the subtexts
and title next to the image have a width of 300px which fixed the issue of trying to
make it look centered. Hopefully that was the way to make it look centered.

Might still need to go back and style the header and quote section if I do not feel
like they are completed.